requirements.txt
opencv-python==4.8.0.76
Pillow==9.4.0
pytesseract==0.3.10